+++ bumper.cpp
@@ -139,7 +139,7 @@ STDMETHODIMP Bumper::InterfaceSupportsErrorInfo(REFIID riid)
       &IID_IBumper,
    };
 
-   for (int i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
+   for (size_t i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
    {
       if (InlineIsEqualGUID(*arr[i],riid))
          return S_OK;
