+++ dispreel.cpp
@@ -388,7 +388,7 @@ STDMETHODIMP DispReel::InterfaceSupportsErrorInfo(REFIID riid)
         &IID_IDispReel,
 	};
 
-	for (int i=0; i<sizeof(arr)/sizeof(arr[0]); ++i)
+	for (size_t i=0; i<sizeof(arr)/sizeof(arr[0]); ++i)
 	{
 		if (InlineIsEqualGUID(*arr[i],riid))
 			return S_OK;
