+++ gate.cpp
@@ -1085,7 +1085,7 @@ STDMETHODIMP Gate::InterfaceSupportsErrorInfo(REFIID riid)
       &IID_IGate,
    };
 
-   for (int i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
+   for (size_t i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
    {
       if (InlineIsEqualGUID(*arr[i],riid))
          return S_OK;
