+++ wintimer.h
@@ -5,6 +5,6 @@
 void wintimer_init();
 
 U32 msec();
-LONGLONG usec();
+unsigned long long usec();
 
 #endif /* !__WINTIMER_H__ */
